+++ airbnb/reservation/src/main/java/com/example/reservation/dto/ReservationDto.java
@@ -1,0 +1,15 @@
+package com.example.reservation.dto;
+
+public class ReservationDto {
+
+    private int rvId;
+
+    private int roomId;
+
+    private String status;
+
+    private int payId;
+
+
+
+}
